requirements_src.txt
notify_py
pynput
PyYAML
screeninfo
structlog
